refactor(storage): name it auto_explain, and keep it in one module (review)
Renames the module and the feature, and moves `with_explain` out of `debug.rs`
to sit with the container settings it pairs with, in
`scheduler_metadata::auto_explain`.

`postgres::auto_explain` is left as the switch: without the feature
`with_explain` is `run` alone, so the call site needs no `cfg` of its own.
